document: add/remove draft-id from session when appropriate
When we save a draft, we want to track the draft-id in "drafts" of the
state object so that it will be known about during the next run of the
text editor. Otherwise we risk deleting the draft file (and loosing the
mapping for the draft-id).

Furthermore, when save the file to the real file, we no longer need to
track the draft as the recent files will give us the ability to have the
file in the sidebar.