session: add EditorSession:auto-save
This adds an auto-save feature to the editor session so that we
can save the session state in the background (without writing
to the actual backing files on storage). If the application
crashes, it can come back with the modified state so the user
has minimal loss to their work.

If the user wants to discard that work, they can use the
"Discard Changes" menu item to release those changes and revert
back to the version found on storage.