document: keep modified bit after saving draft
The saving of the draft can cause the modified bit to be cleared, and we
don't want that. There is a small chance of racing here with saving to
the real file, but our drafts file should be on similar storage or faster
so the chance is relatively low. Either way, a quick follow-up save would
make the race disappear.