document: work around GtkSourceFileSaver for drafts
When saving a draft, GtkSourceView would set gtk_text_buffer_set_modified()
on our behalf. We don't actually want that because we're not saving to the
underlying file, but instead an alternate draft file.

This just works around that by avoiding the save_finish() API for our
particular use-case.

Fixes #25