window: add specialized widget for position label
This allows us to be more specific with sizing requirements so that the
position label doesn't move all over the place when text changes.

Doing this perfectly in all situations is probably not worth the effort,
but this significantly reduces how much it can happen and how often for
most files.

Fixes #28