window: enable DnD for text/uri-list
This has to do a bit of trickery to keep the GtkSourceView from dropping
text/uri-list since text/plain is also provided for the URI. When we
detect there is a text/uri-list, we prevent the drop on the GtkSourceView
and instead let it be caught by the eventbox further up the chain (which
then requests the opening by the session).

I can't figure out why the highlight isn't working on the GtkEventBox
though (in terms of showing a drop target).

Fixes #27