window: add theme selector buttons in menu
This changes from dark-mode (bool) to a style-variant (string) that we
can more easily add to in the future if we end up doing Sepia modes. It
also sets the groundwork for the new selector in the Text Editor mockups
where you can choose light or dark with buttons that look like text
examples.

We still need the icons to set -gtk-icon-source in the CSS, but this gets
the machinery in place.