window: use popover instead of sidebar
This suffles some things around to allow for a popover instead of
a sidebar for the recent items, including search and browse.

Session will return a Page for various operations which might be
useful going forward.

A recents model is now part of the session, making it easier to
bind to the listview.