save-changes: implement mockup for save change dialog
This uses an async/finish pair that can be called from the application
or window actions.

It could definitely use some testing if you're up for it. The mechanics
sort of change a bit than how they were in the past (on purpose) to match
the mockups.

There was a little bit of ambiguity in what to do for drafts that have
not been saved, so I tried to figure out what the mockup meant. My best
guess is that we try to save them in ~/Documents.

If something is not saved (ie: unchecked) I do try to keep the draft
around for the next run to reduce chances of destroying data that the
user may have cared about.