map: add hidden feature for overview map
This allows toggling the GtkSourceMap from a GSettings key. When enabled,
we use a GtkSourceMap instead of a scrollbar.

I should note that on some systems, I see issues with disappearing text
which is probably pointing to a bug in the GL renderer or GtkTextLayout.

Either way, this is off by default and according to designs we do not
display an optnio for it in preferences.