document: tweak line spacing
This makes things look a bit better, but we have to be careful because the
line spacing also affects the overview map. Thankfully, in a way that I
want currently.

However, longer term, we might want to adjust this to use the PangoLayout
if we can affect it to tweak the height.