document: guess content-type after typing first line
This is similar to what we do in Builder so that we can guess content
type after the user types the first line. Emacs also does something
similar.

For common scripting cases, this just works. Put a shebang on the first
line, and it will get picked up. Same for includes on C sources.

Related #24