spellcheck: use sub-iterators to seek further
The goal here is to avoid looking at content from the buffer if we know
that it is either in a region that has 'no-spell-check' context or part
of the TextRegion that we know is already checked.

This tries to employ something similar to how you would do an n-way merge
(very hand-wavy) by applying the forward progress from some sub-iterators
to the others.

There are still some invalidation bugs we need to take care of, but this is
a start in that direction.