spellcheck: plumb support for extra word characters
Based on the language, we might need to extend beyond what GtkTextIter
considers part of a word (from breaking rules). For example, we might get
"re" instead of "we're" in English unless we consider ' as part of the
word.

Further work is needed to introduce this into the forward_word_end() and
backward_word_start().