spellcheck: treat ' as a word char
It would be really nice if we could rely on this elsewhere, but it's going
to be extremely common so for now we'll just hard code it. In Flatpak, I'm
getting 0123456789 for extra_word_chars from enchant, which isn't quite
enough and doesn't match what we get in Fedora with ' and a number of other
characters.