prefs: tweak style-variant when style-scheme changes
This is a minimal fix for #128 to change the style-variant (light/dark)
so that the style scheme will be changed when it is selected.

What we probably want long term is to change the style of the application
theme with the style scheme colors. However, that will require a bit more
work in libadwaita before that is easily doable without the hacks we've
done in various places in Builder.

Currently, for some style schemes you can change the style-variant
manually afterwards from the primary menu, but schemes meant for light/dark
like solarized or Adwaita will still not quite follow this. That is a
known issue and will require us to handle things like mentioned above.