session: raise window when raising page
In case we're raising a page that is already open when the user requests
a file, try to present the window as well so that we hopefully get raised
by the window manager.

We already call gtk_window_present() if the file is not open, but this
ensures that we do it for the raising case too.

Related #169