page: add overlay to jump to line and optional column
Ctrl+I to activate, similar to gedit. Displays an overlay similar to the
search bar which allows inputing line numbers in the form of:

 line
 line:column

both are 1 index based to match the visual display of columns. If no
column is provided, the cursor is placed on the first non-whitespace
character or the end of line if no non-whitespace character is found.

Also added shortcut to Movements section of help-overlay.ui

If/when https://gitlab.gnome.org/GNOME/gtk/-/issues/4315 is fixed, we can
remove the workaround to attach to the GtkText for suppressing non-integral
characters.

Fixes #168