window: remove page from session
If we're going to allow AdwTabView to go through the normal process of
removing pages, then we need to also notify the session of the page
removal so that we don't hold onto it.

This may also improve the situation with a closed tab being obnoxious, such
as in #175, if it ensures the view/buffer are disposed.

Fixes #171