searchbar: simplify search movements
This moves a bunch of our special case keybindings to the search entry
which is where people are actually typing for search movements. We don't
do this for the replace entry because that'd require disambiguating replace
one with replace all.

Additionally, we use Return/Enter to move to the next search result which
feels a bit more natural to users coming from other systems. Using Control
with Return/Enter will move you to the next result and dismiss the search
bar.

You can use Shift combined with Return/Enter to move backwards, and with
Control Return/Enter to move backwards and dismiss the search bar.

Hitting escape without moving to a search result will rubberband you back
to where you started the search. While hitting escape after having moved
to a search result will land you on the current search result.

Fixes #177 and #156