prefs: implement new style scheme chooser
This implements the new style scheme selector design. There may be follow
up work to do on how we want to switch between light and dark variants
based on metadata in the styleschemes, however that can be implemented
elsewhere (in EditorApplication, for example) and probably requires a new
issue specifically for it.

Fixes #178