sourceview: fix line spacing
Originally we had line spacing and it got lost along the way with all the
text buffer changes and CSS changes. This uses a better approach in that
it is part of the CSS generated for the text view directly.

Our original spacing was supposed to be around 1.5x line height, so this
restores that.

The benefit of doing this here without using text tags is that it doesn't
affect the overview map nor does it require adjusting text tags on every
change to the underlying buffer.