session: try to reuse visited time from bookmark
This attempts to use the time from the bookmark so that we don't have to
use the mtime from the underlying file (in most cases).

I don't love that we are using data on the GFile, but it works easy
enough without having to plumb an intermediate structure for the pair.

Addresses comments in #181