window: fix DnD when dropping onto window
We have to duplicate this here and in page because otherwise the text view
will swallow the URI list as a plain UTF-8 string.

So this handles the case of dropping on the header bar or anywhere when a
document is not displayed.

We happend to miss this during the GTK 4 port after DnD was redesigned in
GTK/GDK.

Related #198