sidebarrow: move remove code to open popover
Previously we had the EditorSidebar we were keeping around in case we need
to go back to it from the popover. Since we've pretty much moved on from
that design now, we dropped that code in a previous commit.

That means we can do a simpler design and implement removing rows as an
action on the EditorOpenPopover which simplifies our action muxer quite
a bit.

Fixes #192