sidebaritem: remove use of GPatternSpec
We can do fuzzy matching here instead which is closer to how auto
completion would work in the editor anyway (not that we do that currently).

It also removes the use of GPatternSpec which is a bit tricky to use in a
fast fashion anyway and overkill for our needs here.

Related #192