session: truncate bookmarks to 100
This will truncate the oldest bookmarks beyond 100 so that we don't over
fill our recents list.

For those with a large collection of recent files, this will take two
times to run before the changes are seen. The first run will truncate the
bookmarks when the session is saved (at startup). The second run will
see that truncated list.

Related #192