vim: add support for vim keybindings
This is currently hidden behind the gsettings 'keybindings' key. You may
`gsettings set org.gnome.TextEditor keybindings vim` to enable it, although
in Flatpak you'll need to edit the settings file.

This uses the new GtkSourceVimIMContext along with a command bar widget on
the bottom that is only visible when Vim bindings are enabled.

Use with caution, and file bugs please (or better yet, fix them).