themeselector: add follow button for theme selector
This doesn't actually wire up the "follow" state to track libadwaita
changes, but it does let us see how the styling will work. In particular
my concern is making the border match the mockups.

In the mockups, the border is different for the light and dark half of
the button. For us to do that, we'd likely need to have two buttons and
clip them in a snapshot. An alternative would be to use an SVG.

Instead, this just uses a dark border that looks relatively good in both
light and dark modes.

Related #225