filechooser: try to select a reasonable directory
Instead of always landing on Recent, it's nice to try to be a bit smarter
here. If we have a document open, we can use the directory of that file
first (as it's more likely to be closest).

Otherwise, we can use the directory we last saved to.

Additionally, when saving a new file, we can try to reuse the directory
we last saved within.

Fixes #232