openpopover: tweak spacing for scrollbar
We might need more space for the scrollbar on hover so this tweaks the
margins a bit. I've also added the "card" style class so that lists look
more like they do elsewhere (such as in the preferences pane).

I tried to keep the horizontal padding similar between the top and bottom
but added much more vertical padding on the lower because it felt off
otherwise. YMMV.

Probably needs further design review.

Fixes #233