recoloring: expiriment with recoloring
This can restyle the window based on colors from the GtkSourceStyleScheme.
However, it just uses known styles to do this which is not exactly the
best way to handle it. We need more control over styling from the style
schemes in this case.

My thoughts here are to add additional API to GtkSourceStyleScheme so you
can get named colors. Then the style schemes themselves can export colors
of a given name and if those are found, the scheme is capable for
recoloring.

Related #139