preferences: switch to preferences window
This moves us back to a AdwPreferencesWindow but using a new design. It
does need some more work still, but this is far enough to be generally
useful while we figure out the rest of the tweaks.

It still needs:

 * Preview the custom font (by setting up CSS for the sourceview)
 * Show the proper scheme after light/dark has changed
 * Various other tweaks

But either way good enough to get some feedback.

Related #183