css: lighten source-map a bit
We can't really change foreground colors because they are cached and come
from the GtkSourceBuffer's style scheme. So the only real tweak we can do
to make it drop a bit from focus is to soften it with opacity.

Unfortunately, that affects the slider too since we can't just do this to
the "textview text" selector.