propertiesdialog: add document statistics dialog
This adds a simple properties dialog calculating the statistics of the
document. It is similar to the docinfo plugin of Gedit and in fact shares
the same calculation code (using Pango rather than GtkText*) as those
APIs are significantly faster to calculate.

Some slight changes from the design include:

 * Add "Lines" row. This is mostly useful if we decide to add support for
   selections like the docinfo plugin supports, but also if line numbers
   are not shown or long files.
 * The Location path is clickable to take you to the file within the
   desktop file manager (such as Nautilus in GNOME).
 * An "Update" button is shown in the titlebar if the document changes
   out from under the dialog so the user can refresh the statistics. Given
   how heavy the calculation can be, this is preferred to realtime updates
   of the statistics.

Fixes #264