sourceview: add plumbing to draw search bubbles
Currently, the gtk_text_view_get_iter_location() API does not take into
account the line-height, so this code ported from Builder is not fully
accurate as the line-height grows.

However, it's nice to have a port here so that we can handle that later
once the API is fixed.