document: add an EditorDocument::save signal
This new save signal is emitted before saving the file to the drafts
directory or to the underlying storage. This can be helpful if you only
want to perform an operation when the document is already doing an
expensive operation like writing to disk.

It's a good time to calcuate expensive operations.