settings: allow setting line-height from GSettings
gsettings set org.gnome.TextEditor line-height 2.0 will get you double
spacing line-height. This just uses the CSS line-height attribute (from
newer GTK releases which support it). The default is 1.2 as we've had
since the early days of Text Editor.

We're not currently putting this in the preferences dialog and there is
no plan to, but we can certainly have an override for it in GSettings.

Fixes #285