recoloring: use text fg/bg when possible
If we have values for both fg and bg of text, we should just apply that
to the window/headerbar to ensure we have reasonable contrast as would be
expected by the scheme.

If this isn't good enough contrast, we should blame the style scheme and
fix things there.

Fixes #286