app: avoid restoring full session handling open request
If we get a request to open a file through the GApplicationClass virtual
method, we should avoid restoring the session so the user just sees the
file they requested and not a full session. This should make things a bit
less annoying when opening new files from Nautilus or when launching from
the command-line.

Related #278