subprojects: start on adding subprojects for dependencies
We will need a number of these available as dependencies if we want to be
able to build them on systems other than Flatpak in an automated fashion.

We still need enchant although I'm not quite sure how to go about that
yet (and perhaps we can just do it with dnf instead).