spellcheck: make spell backends optional
This allows us to disable enchant and use a dummy spellcheck backend
instead which can be useful when we need to compile on systems without
enchant (until we get proper backends).

Primarily though, to ease the process of doing CI on systems where getting
it might be annoying or difficult.