spellcheck: use PANGO_UNDERLINE_ERROR_LINE
We can't rely on PANGO_UNDERLINE_ERROR as that may change the baseline
position based on the glyph. This is problematic with mixed language sets
as it comes off very uncoordinated.

With PANGO_UNDERLINE_ERROR_LINE, we also need to set a color since we would
get the text foreground color by default. Plain red (1,0,0) is pretty ugly
so I've just chosen the center red color from the Adwaita palette as our
default red.

Someday, we might want this to come from GtkSourceView, particularly if we
land spellcheck in either GTK or GtkSourceView.

Related to GNOME/gtk#4956
Related to #385