infobar: wrap save button label as well
With the addition of new translations, it appears this button can also
cause the window size to overflow mobile phones screens. Making sure it
has a label on creation allows us to ensure this label can wrap like we
do for other infobar buttons, making this app more suitable to small
displays.