spellcheck: group language selection by group property
This uses the newly added "group" property of EditorSpellLanguageInfo to
group similar languages together in a submenu.

The enchant backend already provides this grouping information via the
uloc_getDisplayLanguage() API.

I do notice some occasional crashes in GTK when shutting down, but that
appears to be unrelated to our code specifically, just that it's not super
great at dealing with menu'ing/action changes in general. We can triage
those upstream as we go though.

Fixes #389