document: disable highlight-matching-brackets during load
This helps ensure we will do less work in GtkSourceBuffer while the
document is loading. Patches have also just landed in GSV to try to do
less work, even if this isn't set for cases where other applications may
not be doing the right thing. But combined, they are even less work.

Related #444