window: dont let AdwTabView steal ctrl+home/end motions
This is a backport from 43 as some platforms appear to have started
shipping Adwaita 1.2 without being aware of the consequences here to
applications.

This does a compile time check for the available API in libadwaita 1.2
so this will need to have a rebuild on a machine that already has a
newer libadwaita-devel available.