session: cleanup remember-recent-files handling
The previous version of this code was not handling a couple details that
weren't obvious at the start, which is that this is running on a thread
and that we don't want to assign to an autoptr.

Given that, the whole "init once" style thing isn't really worth it. So
this instead stores the GSettings in the session struct, if available, and
just uses a bitfield in the save state to determine (on the thread) if
we need to do anything specific.

Fixes #502
Fixes !86