source-view: remove the search bubbles
I'm tired of dealing with these, so just delete them. It still doesn't
fix the primary problem, which is that Adwaita moving to transparent
selections means we can't actually rely on the selection to highlight
the search result, so it's painful to see the difference between active
selection and non-active.