save-changes-dialog: use preferences group for scrolling
The tricky part here is to keep the scrolled area looking like it scrolls
all the way to the edge of the dialog. That requires a bit of copying
what the internal box is doing, which is probably pretty brittle. I'll
need someone with mobile devices to test this out. Otherwise we'll have
to patch the adwaita message dialog (or monkey patch) so that we don't
have the spacing there to begin with.

Fixes #507