view: improve line-wise selections
When selecting lines with ctrl+l or ctrl+d, we need to be more careful
about the presence of CR/LF (multiple end-line characters). Additionally,
it makes sense to avoid perturbing the selection if we detect it is already
something like the line-wise selection we would make.

To make the select_line() code a bit more readable, I've made the extend
line selection code in select_line_action more explicit.

Fixes #434