page: use same directory for save-as except for drafts
We want drafts to continue to use the same directory, but save-as should
use the same directory the file is already in.

This can make it a bit cumbersome to do mass-renames, but that isn't really
a feature we're catering to (although that could be remidied with a
bookmark in the file chooser).

Fixes #603