document: fix etag used for buffer change comparison
We need the appropriate etag so that we can ignore changes that might
happen on delay from the underlying inotify stream. Previously, this was
using the etag of the original file, not the etag of the new file which
was just changed (when doing save as) via
gtk_source_file_saver_save_finish().

Related #615