page: use cancellable for load operations
This uses a cancellable that will emit the cancel signal when the
EditorPage is disposed. It should be disposed when the page is removed
from the AdwTabView.

This ensures that long running load operations from EditorSession, such as
from extremely large files, properly cancels. In turn this ensures that
the forward validation of text within the buffer also cancels.

Fixes #636