window-actions: make language dialog modal
It gets really confusing if multiple of these windows are displayed as
you wont know which dialogs relate to which tab. They were only non-modal
to begin with to allow the user to actually see the content behind the
window.

Instead, just make it modal and hope that the user knows what they want
to change it to anyway.

Fixes #592

Supersedes !136