window: retrain editor when transferring to new window
If we're transferring to a new window, we'll hit this detached_cb. To fix
long running CPU operations in the background we must destroy the
GtkTextView or else it will continue doing forward validation of line text.

But when we are transferring to a new page, we don't want to do that because
we'd become non-functional.

Fixes #643