window: avoid page_detached for destroying view
We're specifically not supposed to do what we were doing, but it was at
the time the only way I found I could get things to break cycles. But with
how things are working now, we can switch this to forcing dispose on the
page and cleaning up template objects.

Fixes #651