map: add  workaround for GtkSourceMap on fractional-scaling
We want this to render faster than it does currently which can be quite
slow on fractional scaling due to anti-aliasing with the border on
possibly merged pixels.

Remove this when GTK's NGL/Vulkan renderer does faster shadows and better
corner clipping w/ rounded borders (which will be needed even if the
shadow part is fixed).