properties-dialog: Don't set the location as a link
When attempting to read the location row, Orca reads the HTML
`<a>` tags. Also, if the file name has spaces or accentuation,
Orca won't read them in a human-readable way.

To fix this, don't set the location as a link, as there are
already the "Open in File Manager" button if the user whishes
to open the file location. Having both the link and the button
is redundant.

Also, now that the location is no longer a link, make the location
row as a property row and use "subtitle-lines=1" to ellipsize the
location, otherwise the row might become too tall if the location
is long.