spellcheck: port to libspelling
This removes all of the original spellcheck code that landed in Text Editor
and uses the libspelling variant instead. Notably, with 0.3.0 that includes
a threaded-spellcheck engine which can do much of the processing off of
the main thread.

We'll need to watch this space closely as we're entering beta right now
for GNOME 47 and I want to ensure this is as stable as possible for that.