build: fix win32.rc build in case the build dir is not a subdir
The icon support was added in !143 but the logic assumed that the
meson build dir is a subdirectory of the source dir, since the
win32.rc gets written to the build dir, but includes the .ico file
relative from the source dir.

In case this is not the case the build fails with something like:
windres.EXE: can't open icon file '../build-aux/win32/org.gnome.TextEditor.ico': No such file or directory

Fix by writing the absolute path to the .ico into the .rc file
at configure time.