view: override extend_selection for source view
In GtkSourceView, it must handle invisible sequences correctly. That can
have drastic performance issues on minified files which we sometimes see
in Text Editor.

Until something more concrete can be handled upstream in GTK and/or
GtkSourceView, mitigate the issue by assuming all text is visible as is
the case in Text Editor.

Fixes: #742