editor-window: Fix zoom buttons UI issues
When zooming in/out, the main menu width keeps changing.
Also, the "Reset Zoom" has a disproportional height
compared to the zoom in/out buttons height.

To fix this, remove valign=center from all zoom buttons
and remove the padding from the "Reset Zoom" button
with CSS. It's the same approach that Ptyxis
uses with its zoom buttons.

Closes https://gitlab.gnome.org/GNOME/gnome-text-editor/-/issues/759