editor-window: Always show a confirmation dialog when discarding changes
In the main menu, the "Discard Changes" can be triggered without
any confirmation. This can lead to data loss if the menu item
is clicked by accident.

Show a confirmation dialog to fix this. Also, since this action
will always require a confirmation, rename the
"page.confirm-discard-changes" to "page.infobar-discard-changes"
to avoid confusion. page.infobar-discard-changes is only used
by EditorInfoBar.