properties-panel: work around subtitle styling with disabled action
The disabled action will cause "Draft" to get dimmed but no the action
representative icon. That is sort of the opposite of what we want. Property
text should be legible and the icon dimmed will indicate that it cannot
have action taken upon it (as well as not highlighting on hover).

Fixes: #768