search-entry: Allow it being recognized by Orca
Also, add labels to it and the replace entry as well.

This makes Orca aware about the wo entries and they're always read when
in the search bar, focusing them has some issues, so the search entry is not
always repored (GTK for some reason does not send the focus event).

Fixes #771.