sidequest: implement various adaptability fixes
This is a huge side-quest to try to fix some issues with small sizing
from #776 that ended up ripping half the window, open-popover, infobars,
property panels, and more apart.

Given the size of this, it will almost certainly have to be GNOME 49
material.

There are some things that aren't quite right yet that need to be done
as a followup to this.

 1. There doesn't seem to be a way in libadwaita to make our Buttom Sheet
    expand all the way to the top (minus some amount) so when you filter
    the search results the bottom sheet will shrink in size. This feels
    extremely wrong from a "widget sizing jumping around" standpoint.

 2. The InfoBars absolutely must be ported away from GtkInfoBar into
    something that will actually work at the smaller scale.

The amount this all got ripped up really does point towards doing a
revamp of the code-base to take all the libadwaita changes into account
from the past 2 years. The code can be modularized much, much better
going forward.