window: shrink open icon when size is limited
Until we can go full-on with the apaptive work we need a bit of a stop
gap to resolve #776.

This handles that (though I've seen it miss by a pixel or two) as a stop
gap measure by swapping the "Open \/" for a simple folder icon once the
size is small eough to really matter.

This generally affects systems with minimize/maximize buttons active.

Fixes: #776