preferences-dialog: add indentation section
This resolves an issue where we lost the ability to tweak some GSettings
which may be necessary by users who've previously changed them as well
as to get expected behavior when editorconfig is unavailable.

This uses all existing strings except one "Indentation" which will be
addressed with followup commits.

Original request for string-freeze break which was not necessary as
all strings are translated and included in the changeset.

https://gitlab.gnome.org/Teams/Translation/Coordination/-/issues/284

Fixes: #782