window: fix bottom sheet to be modal for whole window
Previously, the bottom sheet would not be modal over the top fullscreen
bar and that allowed interacting with the headerbar. We don't want that,
and the style is obviously wrong there.

One weird thing was I had to put the narrow in a fullscreenbox as well
even though we don't care about it simply so the header bar styling would
be the same as you size the window smaller. It would be nice to avoid that
at some point in the future.