search: use wrapper model for search of sidebar items
The goal here is to get some of the search logic outside of the sidebar
model and also avoid use of GtkSorter/GtkFilter machinery. They wont be
able to figure out optimal items-changed emission anyway and our sample
set is extremely small.

So this will do everyhing directly without delays (or incremental updates)
and also gives us a point where we can futuer optimize based on the outcome
of #713 and this change.

One possibility (as noted in the patch) is that we can reduce changed
emissions to be a smaller set by handling the shrinking of search results.
Generally those will continue to be in the same order. When we have that
case we can do incremental drops of items from the listmodel.

This is not implemented yet because we may find that it is not necessary
at all due to the other changes here and previously.

Sorting is also slightly improved to prefer substring runs. This can be a
bit problematic with casefolding but we can give it a shot first and see
how it works out. Realistically we need an actual useful full-text search
instead of this anyway.