Ensure basename property getters return valid UTF-8
g_file_get_basename() is used within getters in the code-base to
retrieve the base filename of a file. However, according to the
documentation, g_file_get_basename() is not guaranteed to return valid
UTF-8 for the purpose of UI display.

Thus it must be made valid if it is invalid if it is to be bound to UI
properties that assume valid UTF-8 is being passed in.

Fixes: #827